requirements.txt
# Install probable requirements
python-crontab==2.7.1